requirements.txt
pillow
PyPDF2
pytesseract
pdfplumber
reportlab
pikepdf
tk